Created delegate for blur checking
